Added some tabs and long integration

+++ src/main/kotlin/vyrek/phasoritenetworks/common/networks/NetworkStatistics.kt
@@ -10,34 +10,23 @@ class NetworkStatistics {
 	enum class EnergyType { IMPORTED, EXPORTED }
 
 	private data class EnergyData(
-		val tickEnergy: MutableList<Int> = mutableListOf(),
-		var transferredEnergy: Int = 0,
-		val energyAverages: ArrayDeque<Double> = ArrayDeque(),
-		var rollingAverage: Int = 0
+		val tickEnergy: MutableList<Long> = mutableListOf(),
+		var transferredEnergy: Long = 0L,
+		val energyAverages: ArrayDeque<Long> = ArrayDeque(),
+		var rollingAverage: Long = 0L
 	)
 
 	private val energyData = mutableMapOf(
 		EnergyType.IMPORTED to EnergyData(),
 		EnergyType.EXPORTED to EnergyData()
 	)
-
-	/**
-	 * Adds the specified energy amount to the statistics for the current tick.
-	 *
-	 * @param energy The amount of energy to be added for the current tick.
-	 * @param type The energy type being added, whether it is input or output energy.
-	 */
-	fun addEnergyTick(energy: Int, type: EnergyType) {
+	
+	fun addEnergyTick(energy: Long, type: EnergyType) {
 		val data = energyData[type]!!
 
 		data.tickEnergy.add(energy)
 	}
 
-	/**
-	 * Called each tick to update network statistics.
-	 *
-	 * @param tickCount Current tick count of the game.
-	 */
 	fun onTick(tickCount: Int) {
 		energyData.forEach { (_, data) ->
 			if (tickCount % (TICKS_PER_SECOND / 4) == 0) {
@@ -49,7 +38,7 @@ class NetworkStatistics {
 			}
 
 			if (tickCount % (TICKS_PER_SECOND * 5) == 0) {
-				data.energyAverages.addFirst(data.rollingAverage.toDouble() / 5)
+				data.energyAverages.addFirst(data.rollingAverage / 5)
 
 				if (data.energyAverages.size >= MAX_ENTRIES) data.energyAverages.removeLast()
 			}
@@ -58,12 +47,7 @@ class NetworkStatistics {
 		}
 	}
 
-	/**
-	 * Get the transferred energy (input/output) of the network.
-	 *
-	 * @param type Whether to get the input or output energy.
-	 */
-	fun getTransferredEnergy(type: EnergyType): Int {
+	fun getTransferredEnergy(type: EnergyType): Long {
 		val data = energyData[type]!!
 
 		return data.transferredEnergy
